+++ examples/configurable/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "basic"
+name = "configurable"
 version = "0.3.0"
 authors = ["Dmitrii Samsonov <blitzarx1@gmail.com>"]
 license = "MIT"
@@ -9,4 +9,6 @@ edition = "2021"
 egui_graphs = { path = "../../" }
 egui = "0.21.0"
 eframe = "0.21.3"
-petgraph = "0.6.3"
+petgraph = "0.6"
+fdg-sim = "0.9.1"
+rand = "0.8.4"
