+++ examples/configurable/README.md
@@ -1,0 +1,8 @@
+# Configurable
+Configurable example where you can toggle settings of the `GraphView` widget and see the result immediately.
+It also contains controls to play with the graph.
+
+## run
+```bash
+cargo run --release -p configurable
+```
