+++ examples/configurable/src/settings.rs
