+++ examples/interactive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "interactive"
-version = "0.2.0"
+version = "0.3.0"
 authors = ["Dmitrii Samsonov <blitzarx1@gmail.com>"]
 license = "MIT"
 edition = "2021"
@@ -10,5 +10,3 @@ egui_graphs = { path = "../../" }
 egui = "0.21.0"
 eframe = "0.21.3"
 petgraph = "0.6"
-fdg-sim = "0.9.1"
-rand = "0.8.4"
