+++ examples/interactive/README.md
@@ -1,6 +1,7 @@
 # Interactive
-Interactive example where you can toggle settings of the `GraphView` widget and see the result immediately.
-It also contains controls to play with the graph.
+Modification of the [basic example](https://github.com/blitzarx1/egui_graph/tree/master/examples/basic) which shows how easy it is to enable interactivity.
+
+Try to drag around and select some nodes.
 
 ## run
 ```bash
